Add check if method has FunctionNameAttribute before other checks
Only methods with a FunctionNameAttribute are relevant for generation.
The generator checks all public types. During the checks (if the necessary attributes are present) it loads dependent assemblies (indirectly - to get the parameters of the methods). During this step an error can occur, because the dependent assembly cannot be loaded.
To avoid such errors, the check for the presence of a FunctionNameAttribute is done before the other checks.