Fix empty warning on build
The output from the generator process is checked and only logged as warning, if the output is not empty.

When a project was built from Visual Studio, the build task logged the process output as warning, even when the output was empty. This resulted in an empty warning in Visual Studio.